fix: Add clipping to time-freq parser.
If a user enters 0Hz, we now translate it to the max value we can
represent with microseconds. If a user enters a value that exceeds 1
Mhz, we clip the periodic value to 1 microsecond. If the user enters a
period that exceeds what will fit in a 32-bit unsigned integer, we clip
it to the max value possible.