refactor: replace error with clipping.
If a value that exceeded a 32-bit integer was specified in a time-freg
field, it would generate a parsing error. Now it gets clipped to
`u32::MAX`. In normal operation, no one ever enters values close to this
max vale, so clipping it removes an error condition no one is checking
for and it still provides an acceptable resolution of large values.