refactor: implement the "immediate" field better.
I never liked how I implemented this; in order to provide compatible
types, I had to use the `char::string()` parsers for "T" and "F".

This commit changes it so that the chained parser returns a `bool`
instead of `String` (by embedding the `value()` parser.) This shortens
the function, removes the ugly `match` statement that contained the same
strings the parser used, and allows us to use `char::char()` parsers for
matching single characters.