:bug: Fix canonical form for event delays.
The DRF documentation indicates the canonical form for delay values is
actually dynamic based on the value. This commit makes the output follow
the document's recommendations.

NOTE: We don't try to reproduce any input using Hz or kHz. The delay
values are converted internally to microseconds. When returning to
canonical form, we don't know if the original form used Hz or a period
so we always use the periodic form.